requirements.txt
beautifulsoup4>=4.12.3
docling>=2.14.0
docling-core>=2.12.1
docling-ibm-models>=3.1.0
docling-parse>=3.0.0
openai>=1.60.1
Pillow>=10.4.0
python-dotenv>=1.0.1
qdrant-client>=1.9.2
requests>=2.32.3
streamlit>=1.37.0
tiktoken>=0.7.0
tqdm>=4.66.4
urllib3>=2.2.2
